chore: rename Utils and install base64 codec statically
