+++ example/src/main/java/cloud/eppo/androidexample/MainActivity.java
@@ -1,6 +1,6 @@
 package cloud.eppo.androidexample;
 
-import static cloud.eppo.android.util.Utils.safeCacheKey;
+import static cloud.eppo.android.util.AndroidUtils.safeCacheKey;
 import static cloud.eppo.androidexample.Constants.INITIAL_FLAG_KEY;
 import static cloud.eppo.androidexample.Constants.INITIAL_SUBJECT_ID;
